RecentsSettings:Make OmniSwitch fragment directly accessible
no need for an extra fragment
