--- res/xml/omniswitch.xml
@@ -1,27 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
- <!-- Copyright (C) 2013 SlimRoms Project
- 
-      Licensed under the Apache License, Version 2.0 (the "License");
-      you may not use this file except in compliance with the License.
-      You may obtain a copy of the License at
- 
-           http://www.apache.org/licenses/LICENSE-2.0
- 
-      Unless required by applicable law or agreed to in writing, software
-      distributed under the License is distributed on an "AS IS" BASIS,
-      WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-      See the License for the specific language governing permissions and
-      limitations under the License.
- -->
- 
- <PreferenceScreen
-     xmlns:android="http://schemas.android.com/apk/res/android"
-     android:title="@string/omniswitch_settings_title">
-
-    <Preference
-            android:key="omniswitch_start_settings"
-            android:title="@string/omniswitch_start_settings_title"
-            android:summary="@string/omniswitch_start_settings_summary"
-            android:persistent="false" />
-
- </PreferenceScreen>
