+++ res/xml/rr_recents.xml
@@ -68,14 +68,13 @@
             android:entryValues="@array/recents_type_title_values"/>
 
         <!-- OmniSwitch -->
-        <PreferenceScreen
+       <Preference
             android:key="omniswitch"
             android:id="@+id/omniswitch"
             android:title="@string/omniswitch_settings_title"
 			android:icon="@drawable/rr_recents_switcher_icon"
-            android:fragment="com.android.settings.rr.OmniSwitch" 
-            android:summary="@string/omniswitch_summary"/>
-
+            android:summary="@string/omniswitch_start_settings_summary"
+            android:persistent="false" />
 
         <Preference
             android:key="slim_recent_panel"
