--- src/com/android/settings/rr/OmniSwitch.java
@@ -1,98 +1,0 @@
-/*
- *  Copyright (C) 2013 The OmniROM Project
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 2 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- *
- */
-
-package com.android.settings.rr;
-
-import com.android.settings.SettingsPreferenceFragment;
-import com.android.settings.R;
-
-import android.app.AlertDialog;
-import android.content.ContentResolver;
-import android.content.Context;
-import android.content.Intent;
-import android.content.DialogInterface;
-import android.os.Bundle;
-import android.support.v7.preference.ListPreference;
-import android.support.v7.preference.Preference;
-import android.support.v7.preference.PreferenceGroup;
-import android.support.v7.preference.PreferenceScreen;
-import android.support.v7.preference.Preference.OnPreferenceChangeListener;
-import android.support.v14.preference.SwitchPreference;
-import android.provider.Settings;
-import android.provider.Settings.SettingNotFoundException;
-import android.util.Log;
-
-import com.android.internal.logging.MetricsProto.MetricsEvent;
-
-public class OmniSwitch extends SettingsPreferenceFragment implements
-        Preference.OnPreferenceChangeListener {
-    private static final String TAG = "OmniSwitch";
-
-    private static final String OMNISWITCH_START_SETTINGS = "omniswitch_start_settings";
-
-    // Package name of the omnniswitch app
-    public static final String OMNISWITCH_PACKAGE_NAME = "org.omnirom.omniswitch";
-    // Intent for launching the omniswitch settings actvity
-    public static Intent INTENT_OMNISWITCH_SETTINGS = new Intent(Intent.ACTION_MAIN)
-            .setClassName(OMNISWITCH_PACKAGE_NAME, OMNISWITCH_PACKAGE_NAME + ".SettingsActivity");
-
-    private Preference mOmniSwitchSettings;
-    private boolean mOmniSwitchInitCalled;
-
-    @Override
-    protected int getMetricsCategory() {
-        return MetricsEvent.RESURRECTED;
-    }
-
-    @Override
-    public void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        addPreferencesFromResource(R.xml.omniswitch);
-
-        PreferenceScreen prefSet = getPreferenceScreen();
-        ContentResolver resolver = getActivity().getContentResolver();
-
-        mOmniSwitchSettings = (Preference)
-                prefSet.findPreference(OMNISWITCH_START_SETTINGS);
-    }
-
-    @Override
-    public boolean onPreferenceTreeClick(Preference preference) {
-        if (preference == mOmniSwitchSettings){
-            startActivity(INTENT_OMNISWITCH_SETTINGS);
-            return true;
-        }
-        return super.onPreferenceTreeClick(preference);
-    }
-
-    public boolean onPreferenceChange(Preference preference, Object objValue) {
-
-        return true;
-    }
-
-    private void openOmniSwitchFirstTimeWarning() {
-        new AlertDialog.Builder(getActivity())
-                .setTitle(getResources().getString(R.string.omniswitch_first_time_title))
-                .setMessage(getResources().getString(R.string.omniswitch_first_time_message))
-                .setNegativeButton(android.R.string.ok, new DialogInterface.OnClickListener() {
-                        public void onClick(DialogInterface dialog, int whichButton) {
-                        }
-                }).show();
-    }
-
-}
